dpdk: Support GCC 10 and custom configuration
GCC 10 is needed for recent archs, because DPDK will take the latest -mtune, so I applied the two concerned patches from upstream on 19.11. It's only false positives -Werror actually.

Also, as I have linking problem, and KNI does not compile in DPDK 19.11 with recent kernel, I propose a way to pass a custom .config file to DPDK. Just putting it in build/dpdk.config and it will be used instead of the default generated .config.