runtime: improvements to preemption support
Cleanups and simplifications to preemption support with the intention
of using periodic yield as a normal scheduling policy.

Also don't mask cede interrupts on delivery. This was done to prevent
stack exhaustion, but it can delay core ceding, which contributes
to tail latency. Hopefully, the IOKernel won't preempt and cede the
same thread repeatedly.