directpath: add queue steering mode
This alternate directpath mode uses a set of always-on RX queues, and
active runtime kthreads are assigned a subset of the RX queues to poll.
While this leads to some inefficiencies, it affords more flexibility to
run on hardware that doesn't support fast flow steering, and is useful
for development and testing.