ksched: fix race condition at thread init
Fix race condition that occurs when the scheduler core wakes a thread
before that thread calls ksched_start(). In this case, the wakeup succeeds
but the thread immediately deschedules itself, leaving the core fully idle.

By adjusting the numbering scheme for the PARKED/UNPARKED enum, we can ensure
that wakeups that occur before ksched_start() is called fail in
try_mark_task_unparked().